Version: 2009-11-30 ksh93t+
Sourced from ast-open-history version 2009-11-30 and:
10b7a377dd87121e7ca2384180e3f5974e81b52c672cc6ee601071dc196fc045  INIT.2009-11-30.tgz

+++ src/lib/libast/vmalloc/vmprivate.c
@@ -247,8 +247,8 @@ int		exact;
 				less = _Vmpagesize;
 			less = (size/less)*less;
 			less = (less/vd->incr)*vd->incr;
-			if(less > 0 && size > less && (size-less) < sizeof(Block_t) )
-				less = less <= vd->incr ? 0 : less - vd->incr;
+			if(less > 0 && size > (size_t)less && (size-(size_t)less) < sizeof(Block_t) )
+				less = (size_t)less <= vd->incr ? 0 : (size_t)less - vd->incr;
 		}
 
 		if(less <= 0 ||
